enhance and separate SlackMember interface, update links to new location
